+++ packages/textarea/src/component.tsx
@@ -12,6 +12,7 @@ import { messages as daMessages } from './locales/da/messages.mjs';
 import { messages as enMessages } from './locales/en/messages.mjs';
 import { messages as fiMessages } from './locales/fi/messages.mjs';
 import { messages as nbMessages } from './locales/nb/messages.mjs';
+import { messages as svMessages } from './locales/sv/messages.mjs';
 import { TextAreaProps } from './props.js';
 import useTextAreaHeight from './useTextAreaHeight.js';
 /**
@@ -35,7 +36,7 @@ export const TextArea = forwardRef<HTMLTextAreaElement, TextAreaProps>((props, f
     ...rest
   } = props;
 
-  activateI18n(enMessages, nbMessages, fiMessages, daMessages);
+  activateI18n(enMessages, nbMessages, fiMessages, daMessages, svMessages);
 
   const id = useId(providedId);
   const ref = useRef<HTMLTextAreaElement | null>(null);